requirements.txt
names==0.3.0
numpy==2.0.1
pandas==2.2.2
plotly==5.23.0
psycopg2==2.9.9
psycopg2_binary==2.9.9
python-dotenv==1.0.1
python_pptx==0.6.23
SQLAlchemy==2.0.31
